+++ plugins/woocommerce-admin/packages/data/CHANGELOG.md
@@ -1,5 +1,6 @@
-# 1.3.2
+# Unreleased
 
+-   Fix error parsing of plugin data package. #7164
 -   Update dependencies
 
 # 1.3.1